pyproject.toml
[project]
name = "review-prompting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "joblib>=1.5.2",
    "langchain>=1.1.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.1",
    "langchain-groq>=1.1.0",
    "langchain-openai>=1.1.0",
    "numpy>=2.3.5",
    "openai>=2.9.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "typing>=3.10.0.0",
]
